This commit was generated by cvs2svn to compensate for changes in r20170, which included commits to RCS files with non-trunk default branches.
git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@20171 728af825-273c-0410-89f9-f7d3b574a069

+++ third/krb5/src/lib/rpc/pmap_clnt.c
@@ -38,6 +38,7 @@ static char sccsid[] = "@(#)pmap_clnt.c 1.37 87/08/11 Copyr 1984 Sun Micro";
  * Copyright (C) 1984, Sun Microsystems, Inc.
  */
 
+#include <unistd.h>
 #include <gssrpc/rpc.h>
 #include <gssrpc/pmap_prot.h>
 #include <gssrpc/pmap_clnt.h>
@@ -59,7 +60,7 @@ pmap_set(program, version, protocol, port)
 	rpc_u_int32 program;
 	rpc_u_int32 version;
 	int protocol;
-	unsigned short port;
+	int port;
 {
 	struct sockaddr_in myaddress;
 	int socket = -1;
